Add conversion script and fix file iteration

+++ callespidf.sh
@@ -18,17 +18,26 @@ fi
 "/Users/ninaherrmann/esp/esp-idf/install.sh"
 . $ESP_EXPORT
 
-idxs=(26)
-for idx in "${idxs[@]}"; do
-  espdlmodelpath="models/espdl/model${idx}.espdl"
-  dest="${PROGRAM_PATH}/model.espdl"
-  echo "Copying $espdlmodelpath -> $dest"
-  rm -f "$dest"
-  cp "$espdlmodelpath" "$dest"
-  # shellcheck disable=SC2164
-  cd "how_to_run_model"
-  idf.py set-target esp32s3
-  idf.py build
-  /Users/ninaherrmann/.espressif/python_env/idf5.5_py3.9_env/bin/python "/Users/ninaherrmann/esp/v5.5.2/esp-idf/tools/idf_size.py" "./build/model_in_flash_rodata.map" > output.txt
-  cd ".."
+for f in *.espdl; do
+    [ -e "$f" ] || continue   # skip literal pattern if no match
+    echo "Processing: $f"  espdlmodelpath="models/espdl/model${idx}.espdl"
+
+    if [[ $f =~ model([0-9]+)_([0-9]+)\.espdl ]]; then
+        idx="${BASH_REMATCH[1]}"
+        seed="${BASH_REMATCH[2]}"
+        echo "First number: $idx"
+        echo "Second number: $seed"
+    fi
+
+    dest="${PROGRAM_PATH}/model.espdl"
+    echo "Copying $espdlmodelpath -> $dest"
+    rm -f "$dest"
+    cp "$espdlmodelpath" "$dest"
+    # shellcheck disable=SC2164
+    cd "how_to_run_model"
+    idf.py set-target esp32s3
+    idf.py build
+    /Users/ninaherrmann/.espressif/python_env/idf5.5_py3.9_env/bin/python "/Users/ninaherrmann/esp/v5.5.2/esp-idf/tools/idf_size.py" "./build/model_in_flash_rodata.map" > output.txt
+    python3 convert_to_csv.py "output.json" --output "memory_results.csv" --idx ${idx} --seed ${seed} --dataset "cifar10"
+    cd ".."
 done
